Implement RuneLite AudioPlayer
